+++ src/com/company/Tecnico.java
@@ -1,8 +1,9 @@
 package com.company;
 
+
 public class Tecnico extends Pessoa {
 
-    public Tecnico(String nome, String email, int cpf, int telefone, double salario){
+    public Tecnico(String nome, String email, long cpf, int telefone, double salario){
         super.nome = nome;
         super.email = email;
         super.cpf = cpf;
